profiler/example: fix zero timestamps in the trace
Problem: the example profiler prints "ts": 0.000000 for every CollApi,
P2pApi and KernelLaunch begin record because exampleProfilerStartEvent
never sets startTs for those three event kinds. KernelCh events on any
channel other than 0 also end with ts 0.0 and StopGpuClk 0: struct
kernelCh declares type as uint8_t while every handle dispatch reads
*(uint64_t*), so the read yields type + 256 * channelId and only channel
0 matches ncclProfileKernelCh. The stop update and the parent refCount
release are skipped for the other channels. Both came in with 2.28.

Solution: stamp startTs for the three API events like every other event
kind, and make kernelCh.type a uint64_t so it agrees with the dispatch
reads (same change as #1973).

Limitations: verified with 2 and 4 ranks on one GPU; multi-GPU, PXN and
CE paths not tested (they do not read these fields differently).

Fixes #1961

+++ plugins/profiler/example/event.h
@@ -72,7 +72,7 @@ struct kernelPhase {
 };
 
 struct kernelCh {
-  uint8_t type;
+  uint64_t type;                    // must be first and 64-bit: handle dispatch reads *(uint64_t*)
   uint8_t channelId;
   struct taskEventBase* parent;
   double startTs;
